chore: extract authorization suite into src/suites/authorization.ts
Part of REAL-1037.
